+++ src/pages/main/Payment.jsx
@@ -1,10 +1,27 @@
 import React, { useState, useEffect } from 'react';
-import { Card, Tabs, Button, DatePicker, Space, Statistic, Row, Col, Table, Typography, Popconfirm } from 'antd';
-import { PlusOutlined, DollarOutlined, ArrowUpOutlined, ArrowDownOutlined, DeleteFilled } from '@ant-design/icons';
+import {
+  Card,
+  Tabs,
+  Button,
+  DatePicker,
+  Space,
+  Statistic,
+  Row,
+  Col,
+  Table,
+  Typography,
+  Popconfirm,
+} from 'antd';
+import {
+  PlusOutlined,
+  DollarOutlined,
+  ArrowUpOutlined,
+  ArrowDownOutlined,
+  DeleteFilled,
+} from '@ant-design/icons';
 import dayjs from 'dayjs';
-import useReports from '../../hooks/useReports';
-import usePayment from '../../hooks/usePayment';
-import useExpenses from '../../hooks/useExpenses';
+import { useQueryClient } from '@tanstack/react-query'; // QueryClient import qilish
+import useFinance from '../../hooks/useFinance'; // Yangi birlashtrilgan hook
 import AddExpenseModal from '../../components/modals/payment/AddExpenseModal';
 import TagUi from '../../components/ui/Tag';
 import toast from 'react-hot-toast';
@@ -21,80 +38,95 @@ const Payment = () => {
   // Sana filtrlash uchun state'lar
   const [filterType, setFilterType] = useState('month');
   const [selectedDate, setSelectedDate] = useState(dayjs());
-  const [selectedRange, setSelectedRange] = useState([dayjs().startOf('month'), dayjs().endOf('month')]);
+  const [selectedRange, setSelectedRange] = useState([
+    dayjs().startOf('month'),
+    dayjs().endOf('month'),
+  ]);
+
+  const queryClient = useQueryClient(); // QueryClient hook
 
   useEffect(() => {
-    const storedBranchId = localStorage.getItem("branchId");
+    const storedBranchId = localStorage.getItem('branchId');
     if (storedBranchId) {
       setBranchId(Number(storedBranchId));
     }
   }, []);
 
-  // Reports hook
-  const {
-    financialSummaryQuery,
-    financialSummaryRangeQuery,
-    refreshFinancialSummary,
-    refreshFinancialSummaryRange
-  } = useReports(branchId);
+  const currentYear = selectedDate?.year() ?? dayjs().year();
+  const currentMonth = (selectedDate?.month() ?? dayjs().month()) + 1;
+  const rangeStartDate = selectedRange?.[0]?.format?.('YYYY-MM-DD') ?? null;
+  const rangeEndDate = selectedRange?.[1]?.format?.('YYYY-MM-DD') ?? null;
 
-  const currentYear = selectedDate.year();
-  const currentMonth = selectedDate.month() + 1;
-  const rangeStartDate = selectedRange[0]?.format('YYYY-MM-DD');
-  const rangeEndDate = selectedRange[1]?.format('YYYY-MM-DD');
-
-  // Payment hook
+  // Yangi birlashtrilgan hook
   const {
+    // Expense operations
+    expensesQuery,
+    deleteExpenseMutation,
+    
+    // Payment operations
     paymentsQuery,
     paymentsByMonthQuery,
     createPaymentMutation,
     deletePaymentMutation,
-    getUnpaidStudentsQuery,
-  } = usePayment({ branchId, year: currentYear, month: currentMonth });
+    unpaidStudentsQuery,
+    
+    // Reports
+    financialSummaryQuery,
+    financialSummaryRangeQuery,
+    refreshFinancialSummary,
+    refreshFinancialSummaryRange,
+    refreshAllData,
+  } = useFinance({ branchId, year: currentYear, month: currentMonth });
 
-  const unpaidStudents = getUnpaidStudentsQuery;
-
-  // Expenses hook
-  const { expensesQuery, deleteExpenseMutation } = useExpenses(branchId);
-
+  // Ma'lumotlarni olish
   const financialSummaryMonth = financialSummaryQuery({ year: currentYear, month: currentMonth });
-  const financialSummaryRange = financialSummaryRangeQuery({ startDate: rangeStartDate, endDate: rangeEndDate });
+  const financialSummaryRange = financialSummaryRangeQuery({ 
+    startDate: rangeStartDate, 
+    endDate: rangeEndDate 
+  });
 
-  const paymentMonth = paymentsByMonthQuery({ branchId, year: currentYear, month: currentMonth }).data;
-
-  const allPayments = paymentsQuery.data;
-  const allExpenses = expensesQuery.data;
+  const paymentsByMonth = paymentsByMonthQuery({ branchId, year: currentYear, month: currentMonth });
+  const allPayments = paymentsQuery?.data ?? [];
+  const allExpenses = expensesQuery?.data ?? [];
+  const unpaidStudents = unpaidStudentsQuery?.data ?? [];
 
   const currentFinancialSummary = filterType === 'range' ? financialSummaryRange : financialSummaryMonth;
 
   const handleAddPayment = (data) => {
+    if (!data) return;
     createPaymentMutation.mutate(data);
   };
 
   // To'lovlar jadvali
   const paymentColumns = [
-    { title: 'Talaba', dataIndex: 'studentName', key: 'studentName' },
-    { title: 'Kurs', dataIndex: 'courseName', key: 'courseName' },
+    { title: "Talaba", dataIndex: "studentName", key: "studentName" },
+    { title: "Kurs", dataIndex: "courseName", key: "courseName" },
     {
-      title: 'Miqdor',
-      dataIndex: 'amount',
-      key: 'amount',
-      render: (amount) => <TagUi>{amount.toLocaleString()} so'm</TagUi>,
+      title: "Miqdor",
+      dataIndex: "amount",
+      key: "amount",
+      render: (amount) =>
+        amount || amount === 0 ? (
+          <TagUi>{Number(amount).toLocaleString()} so'm</TagUi>
+        ) : (
+          <TagUi>Noma'lum</TagUi>
+        ),
     },
     {
-      title: 'Sana',
-      dataIndex: 'date',
-      key: 'date',
-      render: (date) => dayjs(date).format('YYYY-MM-DD HH:mm'),
+      title: "Sana",
+      dataIndex: "date",
+      key: "date",
+      render: (date) => (date ? dayjs(date).format('YYYY-MM-DD HH:mm') : ''),
     },
     {
-      title: 'Izoh',
-      dataIndex: 'description',
-      render: (coment) => coment ? coment.substring(0, 30) : "",
+      title: "Izoh",
+      dataIndex: "description",
+      key: "description",
+      render: (comment) => (comment ? String(comment).substring(0, 30) : ''),
     },
     {
-      title: 'Amallar',
-      key: 'actions',
+      title: "Amallar",
+      key: "actions",
       render: (_, record) => (
         <Popconfirm
           title="Rostdan ham o'chirib tashlaysizmi?"
@@ -107,7 +139,7 @@ const Payment = () => {
             type="primary"
             icon={<DeleteFilled />}
             size="small"
-            loading={deletePaymentMutation.isLoading}
+            loading={deletePaymentMutation?.isLoading}
           />
         </Popconfirm>
       ),
@@ -116,28 +148,33 @@ const Payment = () => {
 
   // Xarajatlar jadvali
   const expenseColumns = [
-    { title: 'Kategoriya', dataIndex: 'category', key: 'category' },
+    { title: "Kategoriya", dataIndex: "category", key: "category" },
     {
-      title: 'Miqdor',
-      dataIndex: 'amount',
-      key: 'amount',
-      render: (amount) => <TagUi color="red">{amount.toLocaleString()} so'm</TagUi>,
+      title: "Miqdor",
+      dataIndex: "amount",
+      key: "amount",
+      render: (amount) =>
+        amount || amount === 0 ? (
+          <TagUi color="red">{Number(amount).toLocaleString()} so'm</TagUi>
+        ) : (
+          <TagUi color="red">Noma'lum</TagUi>
+        ),
     },
     {
-      title: 'Sana',
-      dataIndex: 'date',
-      key: 'date',
-      render: (date) => dayjs(date).format('YYYY-MM-DD HH:mm'),
+      title: "Sana",
+      dataIndex: "date",
+      key: "date",
+      render: (date) => (date ? dayjs(date).format('YYYY-MM-DD HH:mm') : ''),
     },
-    { title: 'Izoh', dataIndex: 'description', key: 'description' },
+    { title: "Izoh", dataIndex: "description", key: "description" },
     {
-      title: 'Amallar',
-      key: 'actions',
+      title: "Amallar",
+      key: "actions",
       render: (_, record) => (
         <Popconfirm
           title="Rostdan ham o'chirib tashlaysizmi?"
           onConfirm={() => {
-            deleteExpenseMutation.mutate(record.id); // Xarajat o'chirish funksiyasi
+            deleteExpenseMutation.mutate(record.id);
           }}
           okText="Ha"
           cancelText="Yo'q"
@@ -147,7 +184,7 @@ const Payment = () => {
             type="primary"
             icon={<DeleteFilled />}
             size="small"
-          // loading={deleteExpenseMutation.isLoading} // Xarajat o'chirish loading
+            loading={deleteExpenseMutation?.isLoading}
           />
         </Popconfirm>
       ),
@@ -157,17 +194,17 @@ const Payment = () => {
   // To'lanmagan talabalar jadvali
   const unpaidStudentsColumns = [
     {
-      title: 'Talaba',
-      key: 'fullName',
-      render: (_, record) => `${record.firstName} ${record.lastName}`,
+      title: "Talaba",
+      key: "fullName",
+      render: (_, record) => `${record?.firstName ?? ''} ${record?.lastName ?? ''}`,
     },
-    { title: 'Guruh', dataIndex: 'groupName', key: 'groupName' },
+    { title: "Guruh", dataIndex: "groupName", key: "groupName" },
     {
-      title: 'Miqdor',
-      dataIndex: 'remainingAmount',
-      key: 'remainingAmount',
+      title: "Miqdor",
+      dataIndex: "remainingAmount",
+      key: "remainingAmount",
       render: (amount) =>
-        amount ? <TagUi>{amount.toLocaleString()} so'm</TagUi> : "Noma'lum",
+        amount || amount === 0 ? <TagUi>{Number(amount).toLocaleString()} so'm</TagUi> : "Noma'lum",
     },
     {
       title: "Amallar",
@@ -177,8 +214,12 @@ const Payment = () => {
           <Popconfirm
             title={
               <div>
-                <div>{record.firstName} {record.lastName}</div>
-                <div>{currentMonth}-oy <strong>{record.groupName}</strong> guruhi</div>
+                <div>
+                  {record?.firstName ?? ''} {record?.lastName ?? ''}
+                </div>
+                <div>
+                  {currentMonth}-oy <strong>{record?.groupName ?? ''}</strong> guruhi
+                </div>
               </div>
             }
             onConfirm={() => {
@@ -195,8 +236,9 @@ const Payment = () => {
             okText="To'lov amalga oshirish"
             cancelText="Yo'q"
           >
-            <Button variant='filled' color='primary'>To'lov Qo'shish</Button>
+            <Button type="primary">To'lov Qo'shish</Button>
           </Popconfirm>
+
           <AddPaymentPopover
             record={record}
             branchId={branchId}
@@ -206,12 +248,37 @@ const Payment = () => {
           />
         </Space>
       ),
-    }
+    },
   ];
 
   const handleDateChange = (date) => {
+    if (!date) return;
+    
+    const newYear = date.year();
+    const newMonth = date.month() + 1;
+    
+    // Eski unpaid-students cache'ini tozalash
+    if (currentYear !== newYear || currentMonth !== newMonth) {
+      // Cache'ni tozalash
+      queryClient?.removeQueries({ 
+        queryKey: ['unpaid-students'], 
+        exact: false 
+      });
+    }
+    
     setSelectedDate(date);
-    setFilterType('day');
+    setFilterType('month');
+  };
+
+  const handleRefresh = () => {
+    if (filterType === 'range') {
+      refreshFinancialSummaryRange();
+    } else {
+      refreshFinancialSummary();
+    }
+    // Barcha ma'lumotlarni yangilash
+    refreshAllData();
+    toast.success('Muvaffaqiyatli yangilandi!');
   };
 
   return (
@@ -219,104 +286,117 @@ const Payment = () => {
       <Title level={2}>Moliyaviy Boshqaruv</Title>
 
       <Card>
-        <Row gutter={16}>
-          <Col span={4} className='justify-center! flex flex-col'>
+        <Row gutter={16} align="middle">
+          <Col span={4} className="flex flex-col justify-center">
             <h6>Oyni Kiriting</h6>
-            <DatePicker
-              picker="month"
-              value={selectedDate}
-              onChange={handleDateChange}
-              allowClear={false}
+            <DatePicker 
+              picker="month" 
+              value={selectedDate} 
+              onChange={handleDateChange} 
+              allowClear={false} 
             />
           </Col>
+
           <Col span={6}>
             <Card bordered={false}>
               <Statistic
                 title="Jami To'lovlar"
-                value={currentFinancialSummary.data?.totalIncome || 0}
+                value={currentFinancialSummary?.data?.totalIncome ?? 0}
                 precision={0}
                 valueStyle={{ color: '#3f8600' }}
                 prefix={<ArrowUpOutlined />}
                 suffix="UZS"
-                loading={currentFinancialSummary.isLoading}
+                loading={currentFinancialSummary?.isLoading}
               />
             </Card>
           </Col>
+
           <Col span={6}>
             <Card bordered={false}>
               <Statistic
                 title="Jami Xarajatlar"
-                value={currentFinancialSummary.data?.totalExpenses || 0}
+                value={currentFinancialSummary?.data?.totalExpenses ?? 0}
                 precision={0}
                 valueStyle={{ color: '#cf1322' }}
                 prefix={<ArrowDownOutlined />}
                 suffix="UZS"
-                loading={currentFinancialSummary.isLoading}
+                loading={currentFinancialSummary?.isLoading}
               />
             </Card>
           </Col>
+
           <Col span={6}>
             <Card bordered={false}>
               <Statistic
                 title="Sof Foyda"
-                value={currentFinancialSummary.data?.netProfit || 0}
+                value={currentFinancialSummary?.data?.netProfit ?? 0}
                 precision={0}
-                valueStyle={{ color: currentFinancialSummary.data?.netProfit >= 0 ? '#3f8600' : '#cf1322' }}
+                valueStyle={{
+                  color:
+                    (currentFinancialSummary?.data?.netProfit ?? 0) >= 0 ? '#3f8600' : '#cf1322',
+                }}
                 prefix={<DollarOutlined />}
                 suffix="UZS"
-                loading={currentFinancialSummary.isLoading}
+                loading={currentFinancialSummary?.isLoading}
               />
             </Card>
           </Col>
-          <Button onClick={() => {
-            if (filterType === 'range') {
-              refreshFinancialSummaryRange();
-              toast.success("Muvaffaqiyatli yangilandi!");
-            } else {
-              refreshFinancialSummary();
-              toast.success("Muvaffaqiyatli yangilandi!");
-            }
-          }}
-            variant="solid" color='cyan'>
-            Yangilash
-          </Button>
 
+          <Col>
+            <Button
+              onClick={handleRefresh}
+              variant="solid"
+              icon={<ArrowUpOutlined rotate={90} />}
+              loading={
+                currentFinancialSummary?.isLoading ||
+                paymentsQuery?.isLoading ||
+                expensesQuery?.isLoading ||
+                unpaidStudentsQuery?.isLoading ||
+                deleteExpenseMutation?.isLoading ||
+                deletePaymentMutation?.isLoading ||
+                createPaymentMutation?.isLoading
+              }
+              color='blue'
+            >
+              Yangilash
+            </Button>
+          </Col>
         </Row>
       </Card>
 
-      <Tabs activeKey={activeTab} className='mt-2' onChange={setActiveTab} type="card">
+      <Tabs activeKey={activeTab} className="mt-2" onChange={(key) => setActiveTab(key)} type="card">
         <TabPane tab="Umumiy To'lovlar" key="1">
-          <h3 className='text-xl font-extrabold'>Barcha To'lovlar</h3>
+          <h3 className="text-xl font-extrabold">Barcha To'lovlar</h3>
           <Table
             columns={paymentColumns}
-            dataSource={allPayments || []}
-            loading={paymentsQuery.isLoading}
+            dataSource={allPayments}
+            loading={paymentsQuery?.isLoading}
             rowKey="id"
             scroll={{ x: 'max-content' }}
           />
         </TabPane>
 
         <TabPane tab="Oylik To'lovlar" key="2">
-          <h3 className='text-xl font-extrabold'>
-            <span className='text-amber-600'>{currentYear} - {currentMonth}</span> uchun To'lovlar
+          <h3 className="text-xl font-extrabold">
+            <span className="text-amber-600">{currentYear} - {currentMonth}</span> uchun To'lovlar
           </h3>
-          <Table
-            columns={paymentColumns}
-            dataSource={paymentMonth || []}
-            rowKey="id"
-            scroll={{ x: 'max-content' }}
+          <Table 
+            columns={paymentColumns} 
+            dataSource={paymentsByMonth?.data ?? []} 
+            loading={paymentsByMonth?.isLoading}
+            rowKey="id" 
+            scroll={{ x: 'max-content' }} 
           />
         </TabPane>
 
         <TabPane tab="To'lanmagan Talabalar" key="3">
-
-          <h3 className='text-xl font-extrabold'>
-            <span className='text-amber-600'>{currentYear} - {currentMonth}</span> uchun To'lanmagan Talabalar
+          <h3 className="text-xl font-extrabold">
+            <span className="text-amber-600">{currentYear} - {currentMonth}</span> uchun To'lanmagan Talabalar
           </h3>
           <Table
             columns={unpaidStudentsColumns}
-            dataSource={unpaidStudents.data}
-            loading={unpaidStudents.isLoading}
+            dataSource={unpaidStudents}
+            loading={unpaidStudentsQuery?.isLoading}
             rowKey="id"
             scroll={{ x: 'max-content' }}
           />
@@ -333,19 +413,20 @@ const Payment = () => {
           </Button>
           <Table
             columns={expenseColumns}
-            dataSource={allExpenses || []}
-            loading={expensesQuery.isLoading}
+            dataSource={allExpenses}
+            loading={expensesQuery?.isLoading}
             rowKey="id"
             scroll={{ x: 'max-content' }}
           />
         </TabPane>
       </Tabs>
-
-      <AddExpenseModal
-        isVisible={isAddExpenseModalVisible}
-        onClose={() => setIsAddExpenseModalVisible(false)}
-        branchId={branchId}
+      {branchId &&
+      <AddExpenseModal 
+        isVisible={isAddExpenseModalVisible} 
+        onClose={() => setIsAddExpenseModalVisible(false)} 
+        branchId={branchId} 
       />
+    }
     </div>
   );
 };
